sdk: Move `Signer` trait to own module
