+++ sdk/src/lib.rs
@@ -40,6 +40,7 @@ pub mod rpc_port;
 pub mod secp256k1_instruction;
 pub mod shred_version;
 pub mod signature;
+pub mod signer;
 pub mod signers;
 pub mod stake_weighted_timestamp;
 pub mod system_transaction;
